Delegate console output to `PluginFactory`
