+++ cli/src/main/kotlin/io/spine/protodata/cli/app/PluginFactory.kt
@@ -26,10 +26,63 @@
 
 package io.spine.protodata.cli.app
 
-import io.spine.protodata.ReflectiveFactory
 import io.spine.protodata.plugin.Plugin
+import io.spine.reflect.Factory
+import io.spine.string.Separator.Companion.nl
+import io.spine.string.pi
+import java.nio.file.Path
+import kotlin.system.exitProcess
 
 /**
- * A reflective factory for [Plugin]s.
+ * The factory for creating [Plugin]s by the names of their classes.
+ *
+ * If a class could not be found, the factory reports the error using the [callback][printError]
+ * passed to the constructor, and then terminates the process with exit code [1][EXIT_CODE].
+ *
+ * @param classLoader
+ *         the loader for the classes of the plugins.
+ * @param classpath
+ *         the user classpath used for finding plugin classes.
+ * @param printError
+ *         the function given to the factory for reporting errors.
  */
-internal class PluginFactory: ReflectiveFactory<Plugin>()
+internal class PluginFactory(
+    classLoader: ClassLoader,
+    private val classpath: List<Path>?,
+    private val printError: (String?) -> Unit
+): Factory<Plugin>(classLoader) {
+
+    fun load(classNames: List<String>): List<Plugin> {
+        return classNames.map { tryCreate(it) }
+    }
+
+    private fun tryCreate(className: String): Plugin {
+        try {
+            return create(className)
+        } catch (e: ClassNotFoundException) {
+            printError(e.stackTraceToString())
+            printAddingToClasspath(className)
+            exitProcess(EXIT_CODE)
+        }
+    }
+
+    private fun printAddingToClasspath(className: String) {
+        printError("Please add the required class `$className` to the user classpath.")
+        if (classpath == null) {
+            printError("No user classpath was provided.")
+            return
+        }
+
+        printError("Provided user classpath:")
+        val cp = classpath
+        val cpStr = cp.joinToString(separator = nl()).pi(indent = " ".repeat(2))
+        printError(cpStr)
+    }
+
+    private companion object {
+        /**
+         * The value for the exit code of the process in case of an error during plugin creation.
+         */
+        const val EXIT_CODE = 1
+    }
+}
